+++ src/helpers/get-from-object-path.ts
@@ -1,4 +1,7 @@
 export const getFromObjectPath = (obj: {}, path: string, defaultValue?: any) => path
   .split('.')
   .reduce((accumulator, currentValue) =>
-    (accumulator && accumulator[currentValue] ? accumulator[currentValue] : (defaultValue || undefined)), obj);
+    (accumulator && (accumulator[currentValue] !== undefined || accumulator[currentValue] !== null)
+      ? accumulator[currentValue]
+      : (defaultValue || undefined))
+    , obj);