+++ src/modules/database-metrics/mongodb/interfaces/cluster-time-signature.interface.ts
@@ -1,0 +1,6 @@
+import { Binary } from 'bson';
+
+export interface IClusterTimeSignature {
+  hash: Binary;
+  keyId: number;
+}