+++ src/modules/database-metrics/mongodb/interfaces/operation-time.interface.ts
@@ -1,0 +1,6 @@
+import { Timestamp } from 'bson';
+
+export interface IOperationTime {
+  ts: Timestamp;
+  t: number;
+}