+++ src/modules/transports/datadog/metrics-map/mongodb-metrics-map.ts
@@ -1,54 +1,54 @@
 export const mongoDbMetricsMap = {
   // read requests per second
-  'mongodb.opcounters.getmoreps': 'opcounters.getmore',
-  'mongodb.opcounters.queryps': 'opcounters.query',
+  'mongodb.opcounters.getmoreps': 'serverStatus.opcounters.getmore',
+  'mongodb.opcounters.queryps': 'serverStatus.opcounters.query',
 
   // write requests per second
-  'mongodb.opcounters.deleteps': 'opcounters.delete',
-  'mongodb.opcounters.insertps': 'opcounters.insert',
-  'mongodb.opcounters.updateps': 'opcounters.update',
+  'mongodb.opcounters.deleteps': 'serverStatus.opcounters.delete',
+  'mongodb.opcounters.insertps': 'serverStatus.opcounters.insert',
+  'mongodb.opcounters.updateps': 'serverStatus.opcounters.update',
 
   // number of clients with reads in progress or queued
-  'mongodb.globallock.activeclients.readers': 'globalLock.activeClients.readers',
+  'mongodb.globallock.activeclients.readers': 'serverStatus.globalLock.activeClients.readers',
 
   // number of clients with writes in progress or queued
-  'mongodb.globallock.activeclients.writers': 'globalLock.activeClients.writers',
+  'mongodb.globallock.activeclients.writers': 'serverStatus.globalLock.activeClients.writers',
 
   // queued read and write requests per second
-  'mongodb.globallock.currentqueue.readers': 'globalLock.currentQueue.readers',
-  'mongodb.globallock.currentqueue.writers': 'globalLock.currentQueue.writers',
+  'mongodb.globallock.currentqueue.readers': 'serverStatus.globalLock.currentQueue.readers',
+  'mongodb.globallock.currentqueue.writers': 'serverStatus.globalLock.currentQueue.writers',
 
   // percentage of client connections used (%)
-  'mongodb.connections.current': 'connections.current',
-  'mongodb.connections.available': 'connections.available',
+  'mongodb.connections.current': 'serverStatus.connections.current',
+  'mongodb.connections.available': 'serverStatus.connections.available',
 
   // max memory usage
-  'mongodb.mem.mapped': 'mem.mapped',
-  'mongodb.mem.resident': 'mem.resident',
-  'mongodb.mem.virtual': 'mem.virtual',
+  'mongodb.mem.mapped': 'serverStatus.mem.mapped',
+  'mongodb.mem.resident': 'serverStatus.mem.resident',
+  'mongodb.mem.virtual': 'serverStatus.mem.virtual',
 
   // number of page faults per host
-  'mongodb.extra_info.page_faultsps': 'extra_info.page_faults',
+  'mongodb.extra_info.page_faultsps': 'serverStatus.extra_info.page_faults',
 
   // data size per replica set
-  'mongodb.stats.datasize': 'metrics.dbStats.dataSize',
+  'mongodb.stats.datasize': 'dbStats.dataSize',
 
   // oplog window per replica set
-  'mongodb.oplog.timediffseconds': '', // ?
+  // 'mongodb.oplog.timediffseconds': '', // TODO: check
 
   // max Replication Lag per replica set
-  // 'mongodb.replset.replicationlag': '', // ? needs to be calculated
+  'mongodb.replset.replicationlag': 'custom.replicationLag',
 
   // number of open cursors per replica set
-  'mongodb.metrics.cursor.open.total': 'metrics.cursor.open.total',
+  'mongodb.cursors.totalopen': 'serverStatus.metrics.cursor.open.total',
 
   // number of timed out cursors per replica set
-  'mongodb.cursors.timedout': 'metrics.cursors.timedOut',
+  'mongodb.cursors.timedout': 'serverStatus.metrics.cursor.timedOut',
 
   // number of asserts per second
-  'mongodb.asserts.msgps': 'asserts,msg',
-  'mongodb.asserts.warningps': 'asserts.warning',
-  'mongodb.asserts.regularps': 'asserts.regular',
-  'mongodb.asserts.userps': 'asserts.user',
-  'mongodb.asserts.rolloversps': 'asserts.rollovers',
+  'mongodb.asserts.msgps': 'serverStatus.asserts.msg',
+  'mongodb.asserts.warningps': 'serverStatus.asserts.warning',
+  'mongodb.asserts.regularps': 'serverStatus.asserts.regular',
+  'mongodb.asserts.userps': 'serverStatus.asserts.user',
+  'mongodb.asserts.rolloversps': 'serverStatus.asserts.rollovers',
 };